Change FlatKey interaction
Fixed that key-inputs could end up in console.
Now keys can only be pressed by keyboard when sitting in chair.
(If complains, use the normal key component).
Changed from direct view line to press, to "just needs to be in front of each other".

Version bump, for this and last commit.