Add debug text to the memory component
Will print the address and the current value of the memory. Additional the character of the servers character mapping. (This mod is dedicated to be used on one specific server, the character set used there will be embedded into this component.
Could be improved by sending polling the character set from the server.
Also makes the font now forced monospace. Eventually the font should be replaced with a real monospace font though.