Optimize custom display
The old custom display was super slow since it updated more often than it had to, and processed each pixel.
Now it updates less and lines without changes get skipped, and all pixel in a line get processed at the same time (with bit-wise operations).
The order of X pegs changed (its reversed now) and the bit order in each byte is also reversed.
Increased major version of this mod, since it is not compatible with the older versions.