CWP: Better calling of BUS feature
Previously it was hacked into CWP by first finishing a CWP action and on start holding the modifier.
However that is better for some other feature, thus the BUS feature should happen in stall mode after the second group was set.
This removes the persistent-ness of it, but makes it much more easy to use.

Version++;
Fixes, that the FirstPersonInteractor was not properly terminated on CWP.