CWP: Use proper routine to wait until next frame
Previously it was a guesswork of waiting for arbitrary time and then joining back on the main thread.
Now a better API is called to execute the routine on the next frame.
It seems that alone the next frame is sufficient for this feature to work. Hence no more repeating.
If that won't work it can be fixed. Now it is way cleaner than before.