CWP: Fix wrongly highlighted and selected pegs for 2D selection
The current 2D peg chooser assumes that the first peg of the first axis is always selected.
However with binary skipping this is not the case (and it won't even less in future).
Now the first peg will actually be properly chosen with an offset according to the second axis.
Further the 2D selected pegs now won't highlight-overwrite pegs which are part of one of the two axes anymore.
This is to be improved further in future.