CWP: Fix expansion and add raycast offset
It is now possible to raycast pegs at the bottom of the pegs, instead of at the top.
Both modes have their advantages depending on the situation.

Also every issue with peg discovering has been fixed. Previously it was possible to bend and reroute the original ray.
Now every peg point will be on that ray. Sadly this requires two raycasts, to get the center point of a peg on the ray, so that distance measurement is precise.