CWP: Fix pattern applying
For the second peg of the second axis the wrong reference point was used, so its position was wrong.
Fail-sounds have been moved out of the internal methods.
The 'onlyOne' argument can now be invoked on all the internal methods too, makes it easier to call.
And the pattern expanding now tries to expand until all pegs are discovered or there are no more.