CWP: Fix SingleWirePlacing mode missing
CWP was mostly developed to run with MWP, from the beginning it was never designed for SWP.
This means that SWP was just a single-group action with 2 pegs.
If single-group connecting was turned off, this won't work anymore.
There still is no dedicated SWP mode, but now there are a bunch of conditions, where only a single wire is placed and the pegs in between are ignored.
Visually this could be improved though.