CWP: Improve handling of ray-cast literal edge case
When the ray only scratches the peg, there is a chance, that it will only hit on the first but not on the second ray-cast.
In such cases the code prints a (bunch of) warning(s) about this. But takes the entry point of the first ray with a small offset.
Previously this would have been handled inconsistently and an exception would break the state of expanding.