ECOMP: Fix breaking IU font of the game
When using the original material of TMP, the UI material of LW UI breaks (with some GPUs).
This can only be prevented by disabling the game object while adding TMP and changing the material.

Also fixed, that the initial memory has no padding before the index numbers.
Removed manual monospacing, as the font is now monospaced.