ECOMP: Don't broadcast memory content on player join
Instead now the client sends its username to the server (as soon as it receives it).
The server will send the memory content to that client, by looking up the connection using the username.

To not collide with file-custom-data deserialization, the username is padded with 1 or 2 extra bytes.

Version increase to 4, as this is now incompatible with previous versions.