FCB: Prevent crash when copying
Not sure what exactly causes this to happen.
But when copying things, the wire-ghosts might not have been initialized, because one of the two peg sides are invisible (for whatever reasons).
Then when updating an outline, the not initialized but invisible wire-ghost is asked to use the pegs to check validity.
-> Crash.

This new patch, will skip wire-ghosts that are invisible, when updating them.