ECOMP: Fix timeout when getting username
The component initialization happens way before the first frame or the server sending the username - which is a huge issue when joining a big world.
Now, the very first setup will be skipped, if there is no playername and the timeout will only count from the first frame onwards. Might still fail, but much less than now.

It would be better to loop setup until the world join has been detected properly and only then apply the timeout.