ECOMP: Remove artifacts from Display
There now is a gap, between the pixel meshes and the main material of the display.
As the faces would clip into each other and cause pixel artifacts.
To not have horrible visuals, the main body of the display is a full cube now.
This means from a (now much longer) distance, the artifacts will be as extreme as vanilla displays on their board.

All in all an improvement for close to mid range between player and display.
-> As future optimization one could use a shader to not render some faces depending on camera position.