RDC: Advanced stacktrace printing & fix double compilation errors
The Harmony patch to print compilation errors was weirdly applied twice. Changed the way its patched which fixed it.

The error screen now shows the full stacktrace.
This removes the need to print the stacktraces to console when opening the error screen.
Stacktraces that Unity captures (and LW prints to console), are now also printed in full to console/logs.

All stacktraces are formatted, so that hashes and line numbers are fully removed - as they are zero anyway.