FCMU: Restore components on world leave
When leaving from multi-player the components won't be restored/updated until next join.
This is an issue for single-player worlds. As these normally have all components available. This can lead to LW trying to force load perfectly fine saves.

Now whenever a world is left, the components will be restored, fixing this issue.