FCMU: Check if server components are installed locally first
The check if components are installed locally has been there for a while.
But a patch sorting components was new, and faulty - it tried to sort components, but did not consider that they might not be installed.

Now it checks if all components are installed first and then sorts and proceeds normally.
For the most part this issue would only occur with an outdated mod version which did not yet provide a component.