Switch all projects to use net7.0
This is mostly and only relevant for the IDE. The game does not care about this.
The server is using NetCore 7, the client still NetFramework 4.8.
Thus for client-only mods this is wrong and could cause compilation errors. But well whatever, I prefer to have it consistent.

ELWAPI made use of both framworks, there always was an error. The net48 feature has been removed. Thus consistent net7 is possible.