EGB: Fix lingering references to WindowWrapper
Instead of the intended actions a reference to the Wrapper's method was added to the closing events. This causes a reference to the WindowWrapper (and thus all of its child components) to remaining outside of the wrapper framework after building.
Meaning Garbage Collection is not able to get rid of it.

Now the actions are just passed through, preventing this reference issue, allowing GC again.