Switch to path variable for referencing LW DLLs
Previously I used (and recommended) symbolic links to the relevant installation directory of Logic World.

However if there is a loop created by symbolic links (game points to mods and mods point to game), then certain lower-tier software like .NET / GitKraken / JetBrains-Rider might fail to resolve the symbolic links and end up spending 100% CPU resolving links until forever. (These programs/software are not low-tier, but not being able to resolve looping symlinks is a reason to be embarrassed - confuses me).

Thus from now on symbolic links are no longer used to point to the game files. Instead MSBuild (environment) variables are used to point towards the game files. These variables can be set up in the used IDE.
No need to set it up outside of it - as the mods do not have to be built. There are surely means to set it up for that as well.

Solves the looping symlinks and keeps the paths in the csproj files cleaner. Might be a bit of a challenge, if the IDE is not automatically adding these paths. Be careful when staging new references.