ELWAPI: Fix VirtualInputPegPool with latest preview
A certain LW developer (me) removed 'StateID' from 'InputPeg' and thought that this makes things easier and more sorted. And thus I removed the Harmony patch (c0f4d9d105619976033ed4b9da98c07c947b2dd8).

Well turns out, this causes the bug which was happening without the patch again, just in the 'ClusterFactory'.
This commit introduces this patch again - just the method to patch was re-introduced somewhere else, so that this patch actually works.
The new method to hook into was added for this purpose in the latest preview (yay, to being a developer).

Recap: The Patch prevents InputPeg-StateID-Update-Mutations from being created for virtual Pegs (would cause issues, as these Pegs do not exist in the world).