EGB+Others: Add feature to set default Y position of custom windows
Sets the default position for new windows. If not changed by a modder, this will be at 750. Besides that if a window has a default size it can be auto-centered.

This commit updates all other GUI mods in this repository to have somewhat matching positions.
Now there is a command to print the Y position of an open EGB window.