WT: Add ability to change outline colors in the settings
The outline colors can now be changed from the LW settings.

Also, there now only is a single outline color for OutputPegs+Wires. As all of them had been some kind of blue anyway.
One could easily re-add this, but which color should the OutputPeg outline be if it is connected to multiple cluster types?