SG: Fix crash when opening loading gui after rejoin
A one line fix!
The action is static, the Instance is static, the callback registration is static.
When registering the callback the 'onMenuShown' method of the original instance is statically referenced. Not the current actual instance.
Thus use lambda, to run the method on the current instance, instead of referencing an (to be deleted) instance method.