Adjust for client now sending mod versions to server
When accessing the mod ID now, one has to map the list of mods - as the data is now in a tuple '(modId, version)'.

Also the PacketIndexOrdering fake-mod has to be injected into the list (instead of the dictionary) now. As LW uses that to create the world join packet.