RC: Initialize VirtualPegPool
I am not why I have not done this before. This was meant to be done.

There is an issue with runtime obtained component addresses when used with VirtualPegPool. It is quite complicated and difficult to reproduce. Anyway, initializing the pool grabs an address from the deleted components. This issue becomes less likely - ofc this is not a fix.

I should address this properly in near future (the next 4 years).