Delete ServerOnlyMods; Update for optional mods/packets
The base game added the 'Optional' Manifest entry. It mainly causes the mod to not be required to be installed on the client anymore while being installed on the server.

This was the core functionality of ServerOnlyMods, thus it is obsolete now. The mod also displayed mod version numbers on the disconnection screen - I transferred that into the base game as well.

Besides that the Optional Packets API is now also obsolete. As the base game now made packet IDs server authoritative and the client will ignore all packets it does not know.