ELWAPI: Fix VirtualPegPool failing to create peg
Specifically when creating a peg during a placement was in progress.
As then LW would increment the highest component address after it grabbed it and grab it before it does anything. A virtual peg and the component would have the same address - bad.
This commit properly fixes this, by reserving component addresses properly at game start instead of at runtime during building actions.